src/gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
redis[asyncio]==5.2.0
openai==1.30.0
prometheus-client==0.21.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
